feat: refactor to use EdgeProfile object
Updated the OpenCommand class to utilize an EdgeProfile object instead of a string for the profile name. This includes changes to the constructor and Name property, as well as command argument handling.

Modified the ProfileList class to support the new EdgeProfile structure, adjusting profile path handling and instantiation of OpenCommand. The EdgeProfile properties are now set to enhance data encapsulation and streamline command invocation.

Updated import statements to include the EdgeProfileCmdPal.Models namespace for the EdgeProfile class definition.